Remove integration tests for Yolov11 and Enable weights from CIv2
