+++ models/demos/yolov11/README.md
@@ -33,7 +33,7 @@ To obtain the perf reports through profiler, please build with following command
 Use the following command to run the model :
 
 ```
-pytest --disable-warnings tests/ttnn/integration_tests/yolov11/test_ttnn_yolov11.py::test_yolov11
+pytest --disable-warnings models/demos/yolov11/tests/pcc/test_ttnn_yolov11.py::test_yolov11
 ```
 
 ### Performant Model with Trace+2CQ
@@ -42,7 +42,7 @@ pytest --disable-warnings tests/ttnn/integration_tests/yolov11/test_ttnn_yolov11
 Use the following command to run the performant Model with Trace+2CQs:
 
 ```
-pytest --disable-warnings models/demos/yolov11/tests/test_e2e_performant.py
+pytest --disable-warnings models/demos/yolov11/tests/perf/test_e2e_performant.py
 ```
 ### Performant Demo with Trace+2CQ
