+++ models/demos/yolov11/common.py
@@ -1,0 +1,28 @@
+# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+
+# SPDX-License-Identifier: Apache-2.0
+
+import os
+
+import torch
+from ultralytics import YOLO
+
+from models.demos.yolov11.reference import yolov11
+
+
+def load_torch_model(model_location_generator=None):
+    torch_model = yolov11.YoloV11()
+    if model_location_generator == None or "TT_GH_CI_INFRA" not in os.environ:
+        model = YOLO("yolo11n.pt")
+        state_dict = {k.replace("model.", "", 1): v for k, v in model.state_dict().items()}
+    else:
+        weights_path = (
+            model_location_generator("vision-models/yolov11n", model_subdir="", download_if_ci_v2=True) / "yolov11n.pt"
+        )
+        state_dict = torch.load(weights_path)
+        state_dict = {k.replace("model.", "", 1): v for k, v in state_dict.items()}
+
+    torch_model.load_state_dict(state_dict)
+    torch_model.eval()
+
+    return torch_model
