+++ models/demos/yolov11/tests/pcc/test_ttnn_yolov11_c2psa.py
@@ -2,34 +2,32 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
-
 import pytest
+
 import ttnn
-import torch
+from models.demos.yolov11.reference.yolov11 import C2PSA as torch_c2psa_block
+from models.demos.yolov11.tt.model_preprocessing import create_yolov11_input_tensors, create_yolov11_model_parameters
+from models.demos.yolov11.tt.ttnn_yolov11_c2psa import TtnnC2PSA as ttnn_c2psa_block
 from tests.ttnn.utils_for_testing import assert_with_pcc
-from models.demos.yolov11.tt.model_preprocessing import (
-    create_yolov11_input_tensors,
-    create_yolov11_model_parameters,
-)
-from models.demos.yolov11.reference.yolov11 import Bottleneck as torch_bottleneck
-from models.demos.yolov11.tt.ttnn_yolov11_bottleneck import TtnnBottleneck as ttnn_bottleneck
 
 
 @pytest.mark.parametrize(
     "in_channel, out_channel, kernel, stride, padding, dilation, groups,fwd_input_shape",
     [
-        ([16, 8], [8, 16], [3, 3], [1, 1], [1, 1], [1, 1], [1, 1], [1, 16, 56, 56]),
-        ([32, 16], [16, 32], [3, 3], [1, 1], [1, 1], [1, 1], [1, 1], [1, 32, 38, 28]),
-        ([32, 32], [32, 32], [3, 3], [1, 1], [1, 1], [1, 1], [1, 1], [1, 32, 14, 14]),
-        ([64, 64], [64, 64], [3, 3], [1, 1], [1, 1], [1, 1], [1, 1], [1, 64, 7, 7]),
-        ([64, 32], [32, 64], [3, 3], [1, 1], [1, 1], [1, 1], [1, 1], [1, 64, 14, 14]),
-        ([32, 16], [16, 32], [3, 3], [1, 1], [1, 1], [1, 1], [1, 1], [1, 32, 28, 28]),
-        ([64, 32], [32, 64], [3, 3], [1, 1], [1, 1], [1, 1], [1, 1], [1, 64, 14, 14]),
-        ([64, 64], [64, 64], [3, 3], [1, 1], [1, 1], [1, 1], [1, 1], [1, 64, 7, 7]),
+        (
+            [256, 256, 128, 128, 128, 128, 256],
+            [256, 256, 256, 128, 128, 256, 128],
+            [1, 1, 1, 1, 3, 1, 1],
+            [1, 1, 1, 1, 1, 1, 1],
+            [0, 0, 0, 0, 1, 0, 0],
+            [1, 1, 1, 1, 1, 1, 1],
+            [1, 1, 1, 1, 128, 1, 1],
+            [1, 256, 7, 7],
+        ),
     ],
 )
 @pytest.mark.parametrize("device_params", [{"l1_small_size": 79104}], indirect=True)
-def test_yolo_v11_bottleneck(
+def test_yolo_v11_c2psa_block(
     device,
     reset_seeds,
     in_channel,
@@ -41,7 +39,7 @@ def test_yolo_v11_bottleneck(
     groups,
     fwd_input_shape,
 ):
-    torch_module = torch_bottleneck(in_channel, out_channel, kernel, stride, padding, dilation, groups)
+    torch_module = torch_c2psa_block(in_channel, out_channel, kernel, stride, padding, dilation, groups)
     torch_module.eval()
     torch_input, ttnn_input = create_yolov11_input_tensors(
         device,
@@ -54,7 +52,7 @@ def test_yolo_v11_bottleneck(
     ttnn_input = ttnn.to_layout(ttnn_input, layout=ttnn.TILE_LAYOUT, memory_config=ttnn.L1_MEMORY_CONFIG)
     torch_output = torch_module(torch_input)
     parameters = create_yolov11_model_parameters(torch_module, torch_input, device=device)
-    ttnn_module = ttnn_bottleneck(device=device, parameter=parameters.conv_args, conv_pt=parameters)
+    ttnn_module = ttnn_c2psa_block(device=device, parameter=parameters.conv_args, conv_pt=parameters)
     ttnn_output = ttnn_module(x=ttnn_input, device=device)
     ttnn_output = ttnn.to_torch(ttnn_output)
     ttnn_output = ttnn_output.permute(0, 3, 1, 2)
