+++ models/demos/yolov11/tests/pcc/test_ttnn_yolov11_c3k2.py
@@ -3,15 +3,12 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import pytest
-import torch
+
 import ttnn
-from tests.ttnn.utils_for_testing import assert_with_pcc
-from models.demos.yolov11.tt.model_preprocessing import (
-    create_yolov11_input_tensors,
-    create_yolov11_model_parameters,
-)
-from models.demos.yolov11.tt.ttnn_yolov11_c3k2 import TtnnC3k2 as ttnn_c3k2
 from models.demos.yolov11.reference.yolov11 import C3k2 as torch_c3k2
+from models.demos.yolov11.tt.model_preprocessing import create_yolov11_input_tensors, create_yolov11_model_parameters
+from models.demos.yolov11.tt.ttnn_yolov11_c3k2 import TtnnC3k2 as ttnn_c3k2
+from tests.ttnn.utils_for_testing import assert_with_pcc
 
 
 @pytest.mark.parametrize(
