+++ models/demos/yolov11/tests/perf/test_e2e_performant.py
@@ -23,7 +23,6 @@ from models.utility_functions import run_for_wormhole_b0
     ],
 )
 @pytest.mark.models_performance_bare_metal
-@pytest.mark.models_performance_virtual_machine
 @run_for_wormhole_b0()
 @pytest.mark.parametrize(
     "device_params", [{"l1_small_size": 24576, "trace_region_size": 6434816, "num_command_queues": 2}], indirect=True
@@ -42,7 +41,7 @@ def test_e2e_performant(
         act_dtype,
         weight_dtype,
         resolution=resolution,
-        model_location_generator=None,
+        model_location_generator=model_location_generator,
     )
     performant_runner._capture_yolov11_trace_2cqs()
     inference_times = []
